pyproject.toml
[project]
name = "agent2agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.42.0",
    "anthropic>=0.57.1",
    "fastapi>=0.116.0",
    "google-adk>=1.5.0",
    "google-cloud-aiplatform>=1.101.0",
    "litellm>=1.74.0.post1",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "pandas-ta",
    "python-dotenv>=1.1.1",
    "setuptools>=80.9.0",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
    "websocket-client>=1.8.0",
]

[tool.uv.sources]
pandas-ta = { git = "https://github.com/superkooks/pandas_ta.git" }
